requirements.txt
requests==2.25.1
xmltodict==0.12.0
tqdm==4.59.0
biopython==1.77
peewee==3.14.4
spacy==3.0.6
scispacy==0.4.0
pandas==1.2.5
PyYAML==5.4.1